refactor: extract question QuestionFooter component
